functional_tests/requirements.txt
pytest==3.10.1
pytest-aiohttp==0.3.0